LayoutTests/http/tests/security/mixedContent/insecure-script-with-secure-cookies-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: secureCookie=yes
CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script-secure-cookies.html requested insecure content from http://127.0.0.1:8080/security/mixedContent/resources/script.js. This content was blocked and must be served over HTTPS.

This test loads a secure iframe that reads secure cookies and then tries to load an insecure script. This should block the insecure script since secure cookies were read.
